+++ neoforge/src/main/java/tocraft/craftedcore/registration/neoforge/KeyBindingRegistryImpl.java
@@ -2,8 +2,11 @@ package tocraft.craftedcore.registration.neoforge;
 
 import com.mojang.logging.LogUtils;
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.Options;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
+import org.apache.commons.lang3.ArrayUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,7 +21,9 @@ public final class KeyBindingRegistryImpl {
         if (!eventAlreadyTriggered) {
             KEY_MAPPINGS.add(keyMapping);
         } else {
-            LogUtils.getLogger().warn("Key Registration Event already run. Couldn't register Key {}", keyMapping.getName());
+            Options options = Minecraft.getInstance().options;
+            options.keyMappings = ArrayUtils.add(options.keyMappings, keyMapping);
+            LogUtils.getLogger().warn("Key Registration Event already run for key {}.", keyMapping.getName());
         }
     }